pyproject.toml
[tool.pdm]

[project]
name = "nonebot-plugin-segbuilder"
version = "0.1.0"
description = "为不同的适配器提供更通用且简易的消息段构建方式"
authors = [
    {name = "Well404", email = "well_404@outlook.com"},
]
dependencies = [
    "anyio>=3.6.2",
    "nonebot-adapter-onebot>=2.2.2",
    "nonebot-adapter-qqguild>=0.2.1",
    "nonebot-adapter-mirai2>=0.0.22",
    "httpx>=0.24.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
[project.optional-dependencies]
dev = [
    "nonebot2[fastapi,websockets]>=2.0.0rc4",
    "pytest-asyncio>=0.21.0",
    "nonebug>=0.3.3",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
